refactor: LottoInfo 클래스의 패키지 변경
domain에서 constant 패키지로 변경
