+++ src/main/java/lotto/util/LottoNumbersGenerator.java
@@ -1,8 +1,8 @@
 package lotto.util;
 
-import static lotto.domain.LottoInfo.END_NUMBER;
-import static lotto.domain.LottoInfo.PICK_COUNT;
-import static lotto.domain.LottoInfo.START_NUMBER;
+import static lotto.constant.LottoInfo.END_NUMBER;
+import static lotto.constant.LottoInfo.PICK_COUNT;
+import static lotto.constant.LottoInfo.START_NUMBER;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
